rtc-websockets/requirements.txt
Django==5.0.6
environs[django]==11.0.0
djangorestframework==3.15.1
django-cors-headers==4.3.1
daphne==4.1.2
channels==4.1.0
redis==5.0.4
channels-redis==4.2.0